Specify width/height on image to avoid layout change
Without this, the layout changes and you see the "Program counter: xxx" text jump around. From the dartdoc on the Image.memory constructor:

Either the [width] and [height] arguments should be specified, or the widget should be placed in a context that sets tight layout constraints.
Otherwise, the image dimensions will change as the image is loaded, which will result in ugly layout changes.

Note: Although this fixes the layout issue, the image still visible flickers because of the asyc loading described above. I'm not sure why this is the case but it seems undesirable and needs another fix!